feat(message bus): got message bus to work with one thread

+++ philo/src/message/formats/message_format_pick_fork.c
@@ -6,7 +6,7 @@
 /*   By: emcnab <marvin@42.fr>                      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/04/11 13:13:33 by emcnab            #+#    #+#             */
-/*   Updated: 2023/04/11 13:14:18 by emcnab           ###   ########.fr       */
+/*   Updated: 2023/04/11 19:19:02 by emcnab           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,10 @@
 
 #include "libft.h"
 
+// BUG: shouls use another function that ft_putnbr_fd so as to correctly handle
+// the entirety of size_t range
 void	message_format_pick_fork(t_s_message *message)
 {
-	ft_putnbr_fd(message->id, STDOUT_FILENO);
+	ft_putnbr_fd((int32_t)message->id, STDOUT_FILENO);
 	ft_putendl_fd(" has picked up a fork", STDOUT_FILENO);
 }
